Add feature to only download blocks needed for file.
Usually all blocks in all pieces that contain the requested file will be
downloaded. This is so that piece validation can be done, which ensures
that the correct data is indeed sent. This can cause a lot more data to
be downloaded than is strictly necessary for the file. For instance,
suppose a torrent with piece length of 32M and a file that is 2M is
desired, then the whole 32M must be downloaded, costing time and bandwidth.
If there is an out of band way to validate the file, eg. MD5 hash, then
there is no benefit to downloaded the extraneous piece data. In this case,
warn if validation is not turned off and turn it off.