Ensure that all piece downloads complete before returning from FilesDownloadManager.get_file_sequential
If we get the last piece number from self.file_pieces, wait for all
remaining piece download tasks to complete.

Each iteration of the while loop in get_file_sequential pops an item from
self.file_pieces, creates a Piece, creates a task for piece.download(),
then checks to see if the task is done. The task is clearly not done because
it will take some time to get the piece data, so then it goes to the next
iteration of the while to repeat for the next piece. The problem is on the
last piece number from self.file_pieces the same thing happens, only now
self.file_downloaded() returns false because self.file_pieces is empty and
the while loop is exited without the piece downloads having been completed.
And since dispatch_manager is never populated either, no piece data is
yielded.