Fix bug where sequential downloader was dropping pieces
The sequential downloader collects finished Piece.download tasks in a
list named task_list. The list task_list is then processed in a for loop,
getting each task and processing it. In the body of the for loop the
task is removed from the task_list using task_list.remove(task). A list
in python should never be modified while it is being iterated over in a
for loop or not all items in the for loop will be processe in the body of
the for loop. This is causing pieces to not be handled and the resulting
download file to not have all piece data written to it, even though all
piece data is downloaded and validated.